chore(db_s3_utils): Rename file to db_utils as we want to use them for more than s3
